+++ src/components/ui/index.ts
@@ -7,3 +7,4 @@ export { default as Button } from "./Button";
 export { default as ProjectCard } from "./ProjectCard";
 export { default as FeaturedProject } from "./FeaturedProject";
 export { default as ExperienceEntry } from "./ExperienceEntry";
+export { default as ResourceCard } from "./ResourceCard";